Improve Hill cipher and add more examples
