+++ BackendReady/CryptoanalysisVigenere.py
@@ -236,11 +236,3 @@ def M_g(frequencies, nPrime, g):
     for i in range(0, 25):
         m_g += (idealDistribution[utils.GetLetter(i)] * frequencies[utils.GetLetter((i + g) % 25)]) / nPrime
     return m_g
-
-"""
-Example
-text = "CHREEVOAHMAERATBIAXXWTNXBEEOPHBSBQMQEQERBWRVXUOAKXAOSXXWEAHBWGJMMQMNKGRFVGXWTRZXWIAKLXFPSKAUTEMNDCMGTSXMXBTUIADNGMGPSRELXNJELXVRVPRTULHDNQWTWDTYGBPHXTFALJHASVBFXNGLLCHRZBWELEKMSJIKNBHWRJGNMGJSGLXFEYPHAGNRBIEQJTAMRVLCRREMNDGLXRRIMGNSNRWCHRQHAEYEVTAQEBBIPEEWEVKAKOEWADREMXMTBHHCHRTKDNVRZCHRCLQOHPWQAIIWXNRMGWOIIFKEE"
-text = utils.preProcessText(text)
-print(GuessKeywordLegth(7, text))
-print(GuessKeyword(5, text))
-"""
